load-tests: split real-token workflows
