+++ bin/load-tester/src/cli.rs
@@ -72,6 +72,11 @@ impl Cli {
 
 /// Load test command arguments.
 #[derive(Args, Clone, Debug)]
+#[command(group(
+    ArgGroup::new("load_mode")
+        .multiple(false)
+        .args(["drain_only", "recover_real_tokens"])
+))]
 struct LoadArgs {
     /// Run continuously until interrupted.
     #[arg(long)]
@@ -90,6 +95,25 @@ struct LoadArgs {
     config: PathBuf,
 }
 
+#[derive(Clone, Copy, Debug, Eq, PartialEq)]
+enum LoadMode {
+    Run,
+    DrainOnly,
+    RecoverRealTokens,
+}
+
+impl LoadMode {
+    const fn from_args(args: &LoadArgs) -> Self {
+        if args.drain_only {
+            Self::DrainOnly
+        } else if args.recover_real_tokens {
+            Self::RecoverRealTokens
+        } else {
+            Self::Run
+        }
+    }
+}
+
 /// Load tester subcommands.
 #[derive(Subcommand, Clone, Debug)]
 enum Command {
@@ -99,6 +123,7 @@ enum Command {
 
 async fn run_load_test(args: LoadArgs) -> Result<()> {
     let mp = LoadTestDisplay::init_tracing();
+    let mode = LoadMode::from_args(&args);
     let config_path = args.config;
 
     if !config_path.exists() {
@@ -125,51 +150,51 @@ async fn run_load_test(args: LoadArgs) -> Result<()> {
 
     let funding_key = TestConfig::funder_key()?;
 
-    if args.drain_only && args.recover_real_tokens {
-        bail!("--drain-only and --recover-real-tokens cannot be used together");
-    }
-
-    // Drain-only mode: recover funds from a previous interrupted run.
-    if args.drain_only {
-        println!("=== Drain-Only Mode ===");
-        println!(
-            "Re-deriving {} accounts from config and draining to funder...",
-            load_config.account_count
-        );
-        let runner = LoadRunner::new(load_config)?;
-        match runner.drain_accounts(funding_key).await {
-            Ok(drained) => println!("Drained {} ETH back to funder.", format_ether(drained)),
-            Err(e) => bail!("drain failed: {e}"),
-        }
-        return Ok(());
-    }
-
-    if args.recover_real_tokens {
-        let Some(real_token_setup) = real_token_setup.as_ref() else {
-            bail!("--recover-real-tokens requires real_token_setup in {}", config_path.display());
-        };
-
-        println!("=== Real-Token Recovery Mode ===");
-        println!(
-            "Re-deriving {} accounts from config and recovering real-token balances to funder...",
-            load_config.account_count
-        );
-        let runner = LoadRunner::new(load_config)?;
-        match runner.recover_real_tokens(real_token_setup).await {
-            Ok(summary) => {
-                println!(
-                    "Recovered {} pair-token raw units and unwrapped {} WETH.",
-                    summary.pair_token_swapped,
-                    format_ether(summary.weth_unwrapped)
-                );
+    match mode {
+        LoadMode::DrainOnly => {
+            println!("=== Drain-Only Mode ===");
+            println!(
+                "Re-deriving {} accounts from config and draining to funder...",
+                load_config.account_count
+            );
+            let runner = LoadRunner::new(load_config)?;
+            match runner.drain_accounts(funding_key).await {
+                Ok(drained) => println!("Drained {} ETH back to funder.", format_ether(drained)),
+                Err(e) => bail!("drain failed: {e}"),
             }
-            Err(e) => bail!("real-token recovery failed: {e}"),
+            return Ok(());
         }
-        match runner.drain_accounts(funding_key).await {
-            Ok(drained) => println!("Drained {} ETH back to funder.", format_ether(drained)),
-            Err(e) => bail!("drain failed: {e}"),
+        LoadMode::RecoverRealTokens => {
+            let Some(real_token_setup) = real_token_setup.as_ref() else {
+                bail!(
+                    "--recover-real-tokens requires real_token_setup in {}",
+                    config_path.display()
+                );
+            };
+
+            println!("=== Real-Token Recovery Mode ===");
+            println!(
+                "Re-deriving {} accounts from config and recovering real-token balances to funder...",
+                load_config.account_count
+            );
+            let runner = LoadRunner::new(load_config)?;
+            match runner.recover_real_tokens(real_token_setup).await {
+                Ok(summary) => {
+                    println!(
+                        "Recovered {} pair-token raw units and unwrapped {} WETH.",
+                        summary.pair_token_swapped,
+                        format_ether(summary.weth_unwrapped)
+                    );
+                }
+                Err(e) => bail!("real-token recovery failed: {e}"),
+            }
+            match runner.drain_accounts(funding_key).await {
+                Ok(drained) => println!("Drained {} ETH back to funder.", format_ether(drained)),
+                Err(e) => bail!("drain failed: {e}"),
+            }
+            return Ok(());
         }
-        return Ok(());
+        LoadMode::Run => {}
     }
 
     println!("=== Base Load Test Runner ===");
