+++ crates/infra/load-tests/src/config/mod.rs
@@ -3,8 +3,13 @@
 mod workload;
 pub use workload::WorkloadConfig;
 
+mod parsing;
+
+mod precompile;
+pub use precompile::PrecompileTarget;
+
+mod real_token;
+pub use real_token::{RealTokenAcquisitionConfig, RealTokenPairTokenConfig, RealTokenSetupConfig};
+
 mod test_config;
-pub use test_config::{
-    OsakaTarget, PrecompileTarget, RealTokenAcquisitionConfig, RealTokenPairTokenConfig,
-    RealTokenSetupConfig, TestConfig, TxTypeConfig, WeightedTxType,
-};
+pub use test_config::{OsakaTarget, TestConfig, TxTypeConfig, WeightedTxType};
