+++ crates/infra/load-tests/src/runner/load_runner.rs
@@ -8,9 +8,7 @@ use std::{
 };
 
 use alloy_network::{Ethereum, EthereumWallet, TransactionBuilder};
-use alloy_primitives::{
-    Address, B256, Bytes, Signed, TxHash, U160, U256, Uint, utils::format_ether,
-};
+use alloy_primitives::{Address, B256, Bytes, TxHash, U256, utils::format_ether};
 use alloy_provider::{Provider, RootProvider};
 use alloy_rpc_types::{BlockNumberOrTag, TransactionRequest};
 use alloy_signer_local::PrivateKeySigner;
@@ -29,20 +27,18 @@ use tokio_util::sync::CancellationToken;
 use tracing::{debug, error, info, instrument, warn};
 
 /// Maximum number of concurrent RPC requests during funding/draining operations.
-const FUNDING_CONCURRENCY: usize = 32;
+pub(super) const FUNDING_CONCURRENCY: usize = 32;
 
 /// Maximum number of TXs to send before waiting for confirmation.
 /// Kept below typical per-sender txpool limits (e.g. reth default is 16) to
 /// avoid "txpool is full" rejections when all TXs originate from one funder.
-const BATCH_SIZE: usize = 16;
-type U24 = Uint<24, 1>;
-type I24 = Signed<24, 1>;
+const FUNDING_BATCH_SIZE: usize = 16;
+const BATCH_SIZE: usize = FUNDING_BATCH_SIZE;
 
 use super::{
     BlockWatcher, DisplaySnapshot, FlashblockWatcher, LoadConfig, LoadTestDisplay, PreparedBatch,
-    PreparedTransaction, QueuedSubmitFailures, RateLimiter, RealTokenAcquisition,
-    RealTokenRecoverySummary, RealTokenSetup, ResultsTracker, SubmissionPipeline, SubmitEvent,
-    TxType,
+    PreparedTransaction, QueuedSubmitFailures, RateLimiter, ResultsTracker, SubmissionPipeline,
+    SubmitEvent, TxType,
 };
 use crate::{
     BaselineError, Result,
@@ -64,17 +60,12 @@ const SUBMIT_WORKER_SHUTDOWN_TIMEOUT: Duration = Duration::from_secs(12);
 const PENDING_CONFIRMATION_TIMEOUT: Duration = Duration::from_secs(60);
 const CONFIRMATION_DRAIN_TIMEOUT: Duration = Duration::from_secs(65);
 const TXPOOL_CLEAR_CONCURRENCY: usize = 64;
-const WETH_DEPOSIT_GAS_LIMIT: u64 = 100_000;
-const WETH_WITHDRAW_GAS_LIMIT: u64 = 45_000;
-const ERC20_APPROVE_GAS_LIMIT: u64 = 65_000;
-const SETUP_SWAP_GAS_LIMIT: u64 = 250_000;
-
 /// Executes load tests by generating and submitting transactions at a target rate.
 pub struct LoadRunner {
-    config: LoadConfig,
+    pub(super) config: LoadConfig,
     config_summary: Option<ConfigSummary>,
-    client: QueryProvider,
-    accounts: AccountPool,
+    pub(super) client: QueryProvider,
+    pub(super) accounts: AccountPool,
     generator: WorkloadGenerator,
     collector: MetricsCollector,
     stop_flag: Arc<AtomicBool>,
@@ -692,19 +683,6 @@ impl LoadRunner {
         routers.into_iter().collect()
     }
 
-    fn collect_real_token_setup_approvals(
-        &self,
-        setup: &RealTokenSetup,
-    ) -> Vec<(Address, Address)> {
-        let mut approvals = HashSet::new();
-        for router in self.collect_swap_routers() {
-            approvals.insert((setup.weth, router));
-            approvals.insert((setup.pair_token.token, router));
-        }
-        approvals.insert((setup.weth, setup.pair_token.acquisition.router()));
-        approvals.into_iter().collect()
-    }
-
     /// Clears pending transactions from all configured txpool nodes for every test sender.
     #[instrument(skip(self), fields(nodes = self.config.txpool_nodes.len(), accounts = self.accounts.len()))]
     pub async fn clear_txpools(&self) -> Result<u64> {
@@ -979,7 +957,7 @@ impl LoadRunner {
         Bytes::from(mintCall { to, amount }.abi_encode())
     }
 
-    fn encode_erc20_balance_of(account: Address) -> Bytes {
+    pub(super) fn encode_erc20_balance_of(account: Address) -> Bytes {
         sol! {
             function balanceOf(address account) external view returns (uint256);
         }
@@ -1443,472 +1421,6 @@ impl LoadRunner {
 
         info!(burned = burn_count, failed = burn_failed, "B-20 teardown complete");
         Ok(())
-    }
-
-    pub async fn setup_real_tokens(&mut self, setup: &RealTokenSetup) -> Result<()> {
-        if self.config.chain_id == 8453 && !setup.allow_chain_id_8453 {
-            return Err(BaselineError::Config(
-                "real-token setup on chain_id 8453 requires allow_chain_id_8453".into(),
-            ));
-        }
-
-        let swap_routers = self.collect_swap_routers();
-        if swap_routers.is_empty() {
-            return Err(BaselineError::Config(
-                "real-token setup requires at least one swap router".into(),
-            ));
-        }
-        let approval_targets = self.collect_real_token_setup_approvals(setup);
-
-        let gas_price = self.client.get_gas_price().await.rpc("get gas price")?;
-        let max_priority_fee = (gas_price / 10).max(1);
-        let max_fee =
-            gas_price.saturating_mul(2).max(max_priority_fee).min(self.config.max_gas_price);
-
-        let account_data: Vec<_> =
-            self.accounts.accounts().iter().map(|a| (a.address, a.signer.clone())).collect();
-        let client = self.client.clone();
-        let primary_submission_rpc = self.config.primary_submission_rpc().clone();
-        let chain_id = self.config.chain_id;
-        let setup = setup.clone();
-        let total_accounts = account_data.len();
-
-        info!(
-            accounts = total_accounts,
-            weth = %setup.weth,
-            pair_token = %setup.pair_token.token,
-            measured_routers = swap_routers.len(),
-            approval_targets = approval_targets.len(),
-            "setting up real-token swap balances"
-        );
-
-        let pb_setup = self.progress_bar(total_accounts as u64, "Setting up real tokens");
-        let setup_futs = account_data.into_iter().map(|(sender, signer)| {
-            let client = client.clone();
-            let primary_submission_rpc = primary_submission_rpc.clone();
-            let setup = setup.clone();
-            let approval_targets = approval_targets.clone();
-            async move {
-                let weth_balance = Self::read_erc20_balance(&client, setup.weth, sender).await?;
-                let pair_balance =
-                    Self::read_erc20_balance(&client, setup.pair_token.token, sender).await?;
-                let needs_pair_token = pair_balance < setup.pair_token.amount_per_sender;
-                let acquisition_amount = if needs_pair_token {
-                    setup.pair_token.acquisition.amount_in()
-                } else {
-                    U256::ZERO
-                };
-                let required_weth =
-                    setup.weth_amount_per_sender.saturating_add(acquisition_amount);
-                let deposit_deficit = required_weth.saturating_sub(weth_balance);
-
-                let mut approvals = Vec::new();
-                for (token, router) in &approval_targets {
-                    let allowance =
-                        Self::read_erc20_allowance(&client, *token, sender, *router).await?;
-                    if allowance < setup.approval_amount {
-                        approvals.push((*token, *router));
-                    }
-                }
-
-                let mut setup_gas_limit = approvals
-                    .len()
-                    .saturating_mul(ERC20_APPROVE_GAS_LIMIT as usize)
-                    as u64;
-                if deposit_deficit > U256::ZERO {
-                    setup_gas_limit = setup_gas_limit.saturating_add(WETH_DEPOSIT_GAS_LIMIT);
-                }
-                if needs_pair_token {
-                    setup_gas_limit = setup_gas_limit.saturating_add(SETUP_SWAP_GAS_LIMIT);
-                }
-
-                let native_balance = client
-                    .get_balance(sender)
-                    .block_id(BlockNumberOrTag::Pending.into())
-                    .await
-                    .rpc("get pending balance")?;
-                let setup_gas_cost = U256::from(setup_gas_limit).saturating_mul(U256::from(max_fee));
-                let total_setup_cost = deposit_deficit.saturating_add(setup_gas_cost);
-                if native_balance < total_setup_cost {
-                    return Err(BaselineError::Transaction(format!(
-                        "sender {} has insufficient balance for real-token setup: has {} ETH, needs {} ETH (deposit {} ETH + setup gas {} ETH)",
-                        sender,
-                        format_ether(native_balance),
-                        format_ether(total_setup_cost),
-                        format_ether(deposit_deficit),
-                        format_ether(setup_gas_cost),
-                    )));
-                }
-
-                let wallet = EthereumWallet::from(signer);
-                let provider = create_wallet_provider(primary_submission_rpc, wallet);
-                let mut nonce = provider
-                    .get_transaction_count(sender)
-                    .pending()
-                    .await
-                    .rpc("get pending transaction count")?;
-                let mut sent = 0usize;
-
-                if deposit_deficit > U256::ZERO {
-                    let tx = TransactionRequest::default()
-                        .with_to(setup.weth)
-                        .with_value(deposit_deficit)
-                        .with_input(Self::encode_weth_deposit())
-                        .with_nonce(nonce)
-                        .with_chain_id(chain_id)
-                        .with_gas_limit(WETH_DEPOSIT_GAS_LIMIT)
-                        .with_max_fee_per_gas(max_fee)
-                        .with_max_priority_fee_per_gas(max_priority_fee);
-                    let pending = provider.send_transaction(tx).await.rpc("send WETH deposit")?;
-                    debug!(
-                        sender = %sender,
-                        tx_hash = %pending.tx_hash(),
-                        amount = %deposit_deficit,
-                        "WETH deposit sent"
-                    );
-                    nonce += 1;
-                    sent += 1;
-                }
-
-                for (token, router) in approvals {
-                    let tx = TransactionRequest::default()
-                        .with_to(token)
-                        .with_input(Self::encode_erc20_approve(router, setup.approval_amount))
-                        .with_nonce(nonce)
-                        .with_chain_id(chain_id)
-                        .with_gas_limit(ERC20_APPROVE_GAS_LIMIT)
-                        .with_max_fee_per_gas(max_fee)
-                        .with_max_priority_fee_per_gas(max_priority_fee);
-                    let pending = provider.send_transaction(tx).await.rpc("send ERC20 approval")?;
-                    debug!(
-                        sender = %sender,
-                        token = %token,
-                        router = %router,
-                        tx_hash = %pending.tx_hash(),
-                        "router approval sent"
-                    );
-                    nonce += 1;
-                    sent += 1;
-                }
-
-                if needs_pair_token {
-                    let (router, data) = Self::encode_pair_token_acquisition(&setup, sender)?;
-                    let tx = TransactionRequest::default()
-                        .with_to(router)
-                        .with_input(data)
-                        .with_nonce(nonce)
-                        .with_chain_id(chain_id)
-                        .with_gas_limit(SETUP_SWAP_GAS_LIMIT)
-                        .with_max_fee_per_gas(max_fee)
-                        .with_max_priority_fee_per_gas(max_priority_fee);
-                    let pending = provider.send_transaction(tx).await.rpc("send setup swap")?;
-                    debug!(
-                        sender = %sender,
-                        router = %router,
-                        tx_hash = %pending.tx_hash(),
-                        "pair-token acquisition swap sent"
-                    );
-                    sent += 1;
-                }
-
-                Ok::<_, BaselineError>((sender, sent))
-            }
-        });
-
-        let setup_results: Vec<_> = stream::iter(setup_futs)
-            .buffer_unordered(FUNDING_CONCURRENCY)
-            .inspect(|_| pb_setup.inc(1))
-            .collect()
-            .await;
-        pb_setup.finish_and_clear();
-
-        let mut sent_total = 0usize;
-        for result in setup_results {
-            let (_, sent) = result?;
-            sent_total += sent;
-        }
-
-        let sender_addresses: Vec<Address> =
-            self.accounts.accounts().iter().map(|a| a.address).collect();
-
-        let mut weth_pending: Vec<_> =
-            sender_addresses.iter().map(|sender| (setup.weth, *sender)).collect();
-        let pb_weth = self.progress_bar(weth_pending.len() as u64, "Waiting for WETH balances");
-        Self::await_token_balances(
-            &self.client,
-            &mut weth_pending,
-            setup.weth_amount_per_sender,
-            &pb_weth,
-        )
-        .await?;
-        pb_weth.finish_and_clear();
-
-        let mut pair_pending: Vec<_> =
-            sender_addresses.iter().map(|sender| (setup.pair_token.token, *sender)).collect();
-        let pb_pair =
-            self.progress_bar(pair_pending.len() as u64, "Waiting for pair-token balances");
-        Self::await_token_balances(
-            &self.client,
-            &mut pair_pending,
-            setup.pair_token.amount_per_sender,
-            &pb_pair,
-        )
-        .await?;
-        pb_pair.finish_and_clear();
-
-        let mut allowance_pending = Vec::new();
-        for sender in &sender_addresses {
-            for (token, router) in &approval_targets {
-                allowance_pending.push((*token, *sender, *router));
-            }
-        }
-        let pb_allowance =
-            self.progress_bar(allowance_pending.len() as u64, "Waiting for router allowances");
-        Self::await_erc20_allowances(
-            &self.client,
-            &mut allowance_pending,
-            setup.approval_amount,
-            &pb_allowance,
-        )
-        .await?;
-        pb_allowance.finish_and_clear();
-
-        self.refresh_sender_state().await?;
-
-        info!(
-            accounts = total_accounts,
-            setup_transactions = sent_total,
-            "real-token setup complete"
-        );
-        Ok(())
-    }
-
-    async fn read_erc20_balance(
-        client: &QueryProvider,
-        token: Address,
-        owner: Address,
-    ) -> Result<U256> {
-        client
-            .call(
-                TransactionRequest::default()
-                    .with_to(token)
-                    .with_input(Self::encode_erc20_balance_of(owner))
-                    .into(),
-            )
-            .await
-            .rpc("eth_call balanceOf")
-            .map(|bytes| U256::from_be_slice(bytes.as_ref()))
-    }
-
-    async fn read_erc20_allowance(
-        client: &QueryProvider,
-        token: Address,
-        owner: Address,
-        spender: Address,
-    ) -> Result<U256> {
-        client
-            .call(
-                TransactionRequest::default()
-                    .with_to(token)
-                    .with_input(Self::encode_erc20_allowance(owner, spender))
-                    .into(),
-            )
-            .await
-            .rpc("eth_call allowance")
-            .map(|bytes| U256::from_be_slice(bytes.as_ref()))
-    }
-
-    fn encode_weth_deposit() -> Bytes {
-        sol! {
-            function deposit() external payable;
-        }
-        Bytes::from(depositCall {}.abi_encode())
-    }
-
-    fn encode_weth_withdraw(amount: U256) -> Bytes {
-        sol! {
-            function withdraw(uint256 wad) external;
-        }
-        Bytes::from(withdrawCall { wad: amount }.abi_encode())
-    }
-
-    fn encode_erc20_approve(spender: Address, amount: U256) -> Bytes {
-        sol! {
-            function approve(address spender, uint256 amount) external returns (bool);
-        }
-        Bytes::from(approveCall { spender, amount }.abi_encode())
-    }
-
-    fn encode_erc20_allowance(owner: Address, spender: Address) -> Bytes {
-        sol! {
-            function allowance(address owner, address spender) external view returns (uint256);
-        }
-        Bytes::from(allowanceCall { owner, spender }.abi_encode())
-    }
-
-    fn encode_pair_token_acquisition(
-        setup: &RealTokenSetup,
-        recipient: Address,
-    ) -> Result<(Address, Bytes)> {
-        match &setup.pair_token.acquisition {
-            RealTokenAcquisition::UniswapV3ExactInput {
-                router,
-                fee,
-                amount_in,
-                min_amount_out,
-            } => {
-                sol! {
-                    interface ISetupUniswapV3Router {
-                        struct ExactInputSingleParams {
-                            address tokenIn;
-                            address tokenOut;
-                            uint24 fee;
-                            address recipient;
-                            uint256 amountIn;
-                            uint256 amountOutMinimum;
-                            uint160 sqrtPriceLimitX96;
-                        }
-
-                        function exactInputSingle(
-                            ExactInputSingleParams calldata params
-                        ) external payable returns (uint256 amountOut);
-                    }
-                }
-                let call = ISetupUniswapV3Router::exactInputSingleCall {
-                    params: ISetupUniswapV3Router::ExactInputSingleParams {
-                        tokenIn: setup.weth,
-                        tokenOut: setup.pair_token.token,
-                        fee: U24::from(*fee),
-                        recipient,
-                        amountIn: *amount_in,
-                        amountOutMinimum: *min_amount_out,
-                        sqrtPriceLimitX96: U160::ZERO,
-                    },
-                };
-                Ok((*router, Bytes::from(call.abi_encode())))
-            }
-            RealTokenAcquisition::AerodromeClExactInput {
-                router,
-                tick_spacing,
-                amount_in,
-                min_amount_out,
-            } => {
-                sol! {
-                    interface ISetupAerodromeClRouter {
-                        struct ExactInputSingleParams {
-                            address tokenIn;
-                            address tokenOut;
-                            int24 tickSpacing;
-                            address recipient;
-                            uint256 deadline;
-                            uint256 amountIn;
-                            uint256 amountOutMinimum;
-                            uint160 sqrtPriceLimitX96;
-                        }
-
-                        function exactInputSingle(
-                            ExactInputSingleParams calldata params
-                        ) external payable returns (uint256 amountOut);
-                    }
-                }
-                let tick_spacing = I24::try_from(*tick_spacing).map_err(|e| {
-                    BaselineError::Config(format!(
-                        "real-token setup tick spacing does not fit i24: {e}"
-                    ))
-                })?;
-                let call = ISetupAerodromeClRouter::exactInputSingleCall {
-                    params: ISetupAerodromeClRouter::ExactInputSingleParams {
-                        tokenIn: setup.weth,
-                        tokenOut: setup.pair_token.token,
-                        tickSpacing: tick_spacing,
-                        recipient,
-                        deadline: U256::from(u64::MAX),
-                        amountIn: *amount_in,
-                        amountOutMinimum: *min_amount_out,
-                        sqrtPriceLimitX96: U160::ZERO,
-                    },
-                };
-                Ok((*router, Bytes::from(call.abi_encode())))
-            }
-        }
-    }
-
-    fn encode_pair_token_recovery(
-        setup: &RealTokenSetup,
-        recipient: Address,
-        amount_in: U256,
-    ) -> Result<(Address, Bytes)> {
-        match &setup.pair_token.acquisition {
-            RealTokenAcquisition::UniswapV3ExactInput { router, fee, .. } => {
-                sol! {
-                    interface IRecoverUniswapV3Router {
-                        struct ExactInputSingleParams {
-                            address tokenIn;
-                            address tokenOut;
-                            uint24 fee;
-                            address recipient;
-                            uint256 amountIn;
-                            uint256 amountOutMinimum;
-                            uint160 sqrtPriceLimitX96;
-                        }
-
-                        function exactInputSingle(
-                            ExactInputSingleParams calldata params
-                        ) external payable returns (uint256 amountOut);
-                    }
-                }
-                let call = IRecoverUniswapV3Router::exactInputSingleCall {
-                    params: IRecoverUniswapV3Router::ExactInputSingleParams {
-                        tokenIn: setup.pair_token.token,
-                        tokenOut: setup.weth,
-                        fee: U24::from(*fee),
-                        recipient,
-                        amountIn: amount_in,
-                        amountOutMinimum: U256::ZERO,
-                        sqrtPriceLimitX96: U160::ZERO,
-                    },
-                };
-                Ok((*router, Bytes::from(call.abi_encode())))
-            }
-            RealTokenAcquisition::AerodromeClExactInput { router, tick_spacing, .. } => {
-                sol! {
-                    interface IRecoverAerodromeClRouter {
-                        struct ExactInputSingleParams {
-                            address tokenIn;
-                            address tokenOut;
-                            int24 tickSpacing;
-                            address recipient;
-                            uint256 deadline;
-                            uint256 amountIn;
-                            uint256 amountOutMinimum;
-                            uint160 sqrtPriceLimitX96;
-                        }
-
-                        function exactInputSingle(
-                            ExactInputSingleParams calldata params
-                        ) external payable returns (uint256 amountOut);
-                    }
-                }
-                let tick_spacing = I24::try_from(*tick_spacing).map_err(|e| {
-                    BaselineError::Config(format!(
-                        "real-token recovery tick spacing does not fit i24: {e}"
-                    ))
-                })?;
-                let call = IRecoverAerodromeClRouter::exactInputSingleCall {
-                    params: IRecoverAerodromeClRouter::ExactInputSingleParams {
-                        tokenIn: setup.pair_token.token,
-                        tokenOut: setup.weth,
-                        tickSpacing: tick_spacing,
-                        recipient,
-                        deadline: U256::from(u64::MAX),
-                        amountIn: amount_in,
-                        amountOutMinimum: U256::ZERO,
-                        sqrtPriceLimitX96: U160::ZERO,
-                    },
-                };
-                Ok((*router, Bytes::from(call.abi_encode())))
-            }
-        }
     }
 
     /// Runs the load test and returns metrics summary.
@@ -2429,182 +1941,6 @@ impl LoadRunner {
         }
     }
 
-    /// Recovers real-token balances by swapping the configured pair token back
-    /// into WETH, unwrapping WETH to native ETH, and leaving native drain to
-    /// [`Self::drain_accounts`].
-    pub async fn recover_real_tokens(
-        &self,
-        setup: &RealTokenSetup,
-    ) -> Result<RealTokenRecoverySummary> {
-        let client = self.client.clone();
-        let primary_submission_rpc = self.config.primary_submission_rpc().clone();
-        let chain_id = self.config.chain_id;
-
-        let gas_price = client.get_gas_price().await.rpc("get gas price")?;
-        let max_priority_fee = (gas_price / 10).max(1);
-        let max_fee =
-            gas_price.saturating_mul(2).max(max_priority_fee).min(self.config.max_gas_price);
-
-        let account_data: Vec<_> =
-            self.accounts.accounts().iter().map(|a| (a.address, a.signer.clone())).collect();
-        let total_accounts = account_data.len();
-        let pb_recover = self.progress_bar(total_accounts as u64, "Recovering real tokens");
-
-        let recover_futs: Vec<_> = account_data
-            .into_iter()
-            .map(|(sender, signer)| {
-                let client = client.clone();
-                let primary_submission_rpc = primary_submission_rpc.clone();
-                let setup = setup.clone();
-                async move {
-                    let mut summary = RealTokenRecoverySummary::default();
-                    let wallet = EthereumWallet::from(signer);
-                    let provider = create_wallet_provider(primary_submission_rpc, wallet);
-                    let mut nonce = provider
-                        .get_transaction_count(sender)
-                        .pending()
-                        .await
-                        .rpc("get pending transaction count")?;
-
-                    let pair_balance =
-                        Self::read_erc20_balance(&client, setup.pair_token.token, sender).await?;
-                    if pair_balance > U256::ZERO {
-                        let recovery_router = setup.pair_token.acquisition.router();
-                        let allowance = Self::read_erc20_allowance(
-                            &client,
-                            setup.pair_token.token,
-                            sender,
-                            recovery_router,
-                        )
-                        .await?;
-
-                        if allowance < pair_balance {
-                            let approval_amount = if setup.approval_amount > pair_balance {
-                                setup.approval_amount
-                            } else {
-                                pair_balance
-                            };
-                            let tx = TransactionRequest::default()
-                                .with_to(setup.pair_token.token)
-                                .with_input(Self::encode_erc20_approve(
-                                    recovery_router,
-                                    approval_amount,
-                                ))
-                                .with_nonce(nonce)
-                                .with_chain_id(chain_id)
-                                .with_gas_limit(ERC20_APPROVE_GAS_LIMIT)
-                                .with_max_fee_per_gas(max_fee)
-                                .with_max_priority_fee_per_gas(max_priority_fee);
-                            let pending =
-                                provider.send_transaction(tx).await.rpc("send pair approval")?;
-                            debug!(
-                                sender = %sender,
-                                token = %setup.pair_token.token,
-                                router = %recovery_router,
-                                tx_hash = %pending.tx_hash(),
-                                "recovery pair-token approval sent"
-                            );
-                            let receipt =
-                                pending.get_receipt().await.rpc("confirm pair approval")?;
-                            if !receipt.status() {
-                                return Err(BaselineError::Transaction(format!(
-                                    "pair-token approval reverted for sender {sender}"
-                                )));
-                            }
-                            nonce += 1;
-                        }
-
-                        let (router, data) =
-                            Self::encode_pair_token_recovery(&setup, sender, pair_balance)?;
-                        let tx = TransactionRequest::default()
-                            .with_to(router)
-                            .with_input(data)
-                            .with_nonce(nonce)
-                            .with_chain_id(chain_id)
-                            .with_gas_limit(SETUP_SWAP_GAS_LIMIT)
-                            .with_max_fee_per_gas(max_fee)
-                            .with_max_priority_fee_per_gas(max_priority_fee);
-                        let pending = provider
-                            .send_transaction(tx)
-                            .await
-                            .rpc("send pair-token recovery swap")?;
-                        debug!(
-                            sender = %sender,
-                            router = %router,
-                            amount = %pair_balance,
-                            tx_hash = %pending.tx_hash(),
-                            "pair-token recovery swap sent"
-                        );
-                        let receipt =
-                            pending.get_receipt().await.rpc("confirm pair-token recovery swap")?;
-                        if !receipt.status() {
-                            return Err(BaselineError::Transaction(format!(
-                                "pair-token recovery swap reverted for sender {sender}"
-                            )));
-                        }
-                        nonce += 1;
-                        summary.pair_token_swapped =
-                            summary.pair_token_swapped.saturating_add(pair_balance);
-                    }
-
-                    let weth_balance =
-                        Self::read_erc20_balance(&client, setup.weth, sender).await?;
-                    if weth_balance > U256::ZERO {
-                        let tx = TransactionRequest::default()
-                            .with_to(setup.weth)
-                            .with_input(Self::encode_weth_withdraw(weth_balance))
-                            .with_nonce(nonce)
-                            .with_chain_id(chain_id)
-                            .with_gas_limit(WETH_WITHDRAW_GAS_LIMIT)
-                            .with_max_fee_per_gas(max_fee)
-                            .with_max_priority_fee_per_gas(max_priority_fee);
-                        let pending =
-                            provider.send_transaction(tx).await.rpc("send WETH withdraw")?;
-                        debug!(
-                            sender = %sender,
-                            amount = %weth_balance,
-                            tx_hash = %pending.tx_hash(),
-                            "recovery WETH withdraw sent"
-                        );
-                        let receipt = pending.get_receipt().await.rpc("confirm WETH withdraw")?;
-                        if !receipt.status() {
-                            return Err(BaselineError::Transaction(format!(
-                                "WETH withdraw reverted for sender {sender}"
-                            )));
-                        }
-                        summary.weth_unwrapped =
-                            summary.weth_unwrapped.saturating_add(weth_balance);
-                    }
-
-                    Ok::<_, BaselineError>(summary)
-                }
-            })
-            .collect();
-
-        let recover_results: Vec<_> = stream::iter(recover_futs)
-            .buffer_unordered(FUNDING_CONCURRENCY)
-            .inspect(|_| pb_recover.inc(1))
-            .collect()
-            .await;
-        pb_recover.finish_and_clear();
-
-        let mut summary = RealTokenRecoverySummary::default();
-        for result in recover_results {
-            let account_summary = result?;
-            summary.pair_token_swapped =
-                summary.pair_token_swapped.saturating_add(account_summary.pair_token_swapped);
-            summary.weth_unwrapped =
-                summary.weth_unwrapped.saturating_add(account_summary.weth_unwrapped);
-        }
-
-        info!(
-            pair_token_swapped = %summary.pair_token_swapped,
-            weth_unwrapped = %summary.weth_unwrapped,
-            "real-token recovery complete"
-        );
-        Ok(summary)
-    }
-
     /// Drains all test account balances back to the funder address.
     ///
     /// Each account sends its entire balance minus gas costs back to the funder.
@@ -2728,7 +2064,7 @@ impl LoadRunner {
         Ok(total_drained)
     }
 
-    fn progress_bar(&self, total: u64, prefix: &str) -> ProgressBar {
+    pub(super) fn progress_bar(&self, total: u64, prefix: &str) -> ProgressBar {
         if self.snapshot_tx.is_some() {
             return ProgressBar::hidden();
         }
@@ -2788,7 +2124,7 @@ impl LoadRunner {
     }
 
     /// Waits for token balances to reach a target after mint/distribution transactions.
-    async fn await_token_balances(
+    pub(super) async fn await_token_balances(
         client: &QueryProvider,
         pending_accounts: &mut Vec<(Address, Address)>,
         target_balance: U256,
@@ -2841,57 +2177,7 @@ impl LoadRunner {
         Ok(settled)
     }
 
-    /// Waits for ERC20 allowances to reach a target.
-    async fn await_erc20_allowances(
-        client: &QueryProvider,
-        pending_allowances: &mut Vec<(Address, Address, Address)>,
-        target_allowance: U256,
-        pb: &ProgressBar,
-    ) -> Result<usize> {
-        let timeout = Duration::from_secs(60);
-        let poll_interval = Duration::from_millis(500);
-        let start = Instant::now();
-        let mut settled = 0usize;
-
-        while !pending_allowances.is_empty() && start.elapsed() < timeout {
-            tokio::time::sleep(poll_interval).await;
-
-            let mut still_pending = Vec::new();
-            for (token, owner, spender) in pending_allowances.drain(..) {
-                match Self::read_erc20_allowance(client, token, owner, spender).await {
-                    Ok(allowance) if allowance >= target_allowance => {
-                        debug!(token = %token, owner = %owner, spender = %spender, "allowance settled");
-                        settled += 1;
-                        pb.inc(1);
-                    }
-                    Ok(_) => {
-                        still_pending.push((token, owner, spender));
-                    }
-                    Err(e) => {
-                        warn!(
-                            token = %token,
-                            owner = %owner,
-                            spender = %spender,
-                            error = %e,
-                            "failed to check allowance"
-                        );
-                        still_pending.push((token, owner, spender));
-                    }
-                }
-            }
-            *pending_allowances = still_pending;
-        }
-
-        if !pending_allowances.is_empty() {
-            return Err(BaselineError::Transaction(format!(
-                "allowances did not reach target within timeout: {pending_allowances:?}"
-            )));
-        }
-
-        Ok(settled)
-    }
-
-    async fn refresh_sender_state(&mut self) -> Result<()> {
+    pub(super) async fn refresh_sender_state(&mut self) -> Result<()> {
         let total_accounts = self.accounts.len();
         let client = self.client.clone();
         let pb_refresh = self.progress_bar(total_accounts as u64, "Refreshing account state");
