+++ crates/infra/load-tests/src/runner/mod.rs
@@ -36,3 +36,5 @@ pub use status::{DisplaySnapshot, LoadTestDisplay};
 
 mod load_runner;
 pub use load_runner::LoadRunner;
+
+mod real_tokens;
